requirements.txt
keyboard
phonemizer
torch
transformers
scipy
munch
sounddevice
python-multipart
soundfile
pydantic
requests
python-dotenv
numpy
pyaudio
pyannote.audio
torch_audiomentations
pydantic_settings
faster-whisper
huggingface_hub
kokoro
